+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to this project will be documented in this file, following t
 - interpret 'CONECT 1529 1528 1528' as double bond in pdb files
 - `side` Buffer/Representation parameter changed: THREE.FrontSide => "front", THREE.BackSide => "back", THREE.DoubleSide => "double"
 - support for negative resno values in selections: "-5:A", "-12--8", "-12-0"
+- support for chemical component type (available in mmtf) for determining molecule type
 
 ### Removed
 - python-based build scripts